+++ jsk_pcl_ros/package.xml
@@ -44,6 +44,7 @@
   <build_depend>jsk_topic_tools</build_depend>
   <build_depend>roscpp_tutorials</build_depend>
   <build_depend>jsk_footstep_msgs</build_depend>
+  <build_depend>jsk_recognition_utils</build_depend>
   <run_depend>roscpp_tutorials</run_depend>
   <run_depend>jsk_topic_tools</run_depend>
