+++ jsk_perception/jsk_perception_nodelets.xml
@@ -39,6 +39,10 @@
          type="jsk_perception::PolygonArrayToLabelImage"
          base_class_type="nodelet::Nodelet">
   </class>
+  <class name="jsk_perception/TabletopColorDifferenceLikelihood"
+         type="jsk_perception::TabletopColorDifferenceLikelihood"
+         base_class_type="nodelet::Nodelet">
+  </class>
   <class name="jsk_perception/ErodeMaskImage"
          type="jsk_perception::Erode"
          base_class_type="nodelet::Nodelet">
