+++ jsk_perception/package.xml
@@ -44,6 +44,7 @@
   <build_depend>jsk_recognition_msgs</build_depend>
   <build_depend>posedetection_msgs</build_depend>
   <build_depend>roseus</build_depend>
+  <build_depend>jsk_recognition_utils</build_depend>
   <run_depend>rospy</run_depend>
   <!-- <run_depend>roseus</run_depend> -->
   <run_depend>std_msgs</run_depend>
@@ -73,6 +74,7 @@
   <run_depend>jsk_recognition_msgs</run_depend>
   <run_depend>posedetection_msgs</run_depend>
   <run_depend>roseus</run_depend>
+  <run_depend>jsk_recognition_utils</run_depend>
   <export>
     <nodelet plugin="${prefix}/jsk_perception_nodelets.xml"/>
   </export>
