+++ jsk_recognition_utils/include/jsk_recognition_utils/pcl/ear_clipping_patched.h
