+++ jsk_recognition_utils/src/random_util.cpp
@@ -33,51 +33,34 @@
  *  POSSIBILITY OF SUCH DAMAGE.
  *********************************************************************/
 
-#ifndef JSK_PCL_ROS_SPINDLE_LASER_SENSOR_H_
-#define JSK_PCL_ROS_SPINDLE_LASER_SENSOR_H_
+#include "jsk_recognition_utils/random_util.h"
+#include <iostream>
 
-#include "jsk_pcl_ros/sensor_model/pointcloud_sensor_model.h"
-
-namespace jsk_pcl_ros
+namespace jsk_recognition_utils
 {
-  class SpindleLaserSensor: public PointCloudSensorModel
+  double randomGaussian(double mean, double var, boost::mt19937& gen)
   {
-  public:
-    typedef boost::shared_ptr<SpindleLaserSensor> Ptr;
-    
-    SpindleLaserSensor(const double min_angle, const double max_angle,
-                       const size_t point_sample):
-      min_angle_(min_angle), max_angle_(max_angle),
-      point_sample_(point_sample) { }
-    
-    virtual void setSpindleVelocity(const double velocity)
-    {
-      spindle_velocity_ = spindle_velocity;
+    if (var == 0.0) {
+      return mean;
     }
-
-    /**
-     * @brief
-     * Return the expected number of points according to distance and area.
-     * it is calculated according to:
-     * \frac{N}{2 \pi \Delta \phi}\frac{1}{r^2}s
-     * \Delta \phi = \frac{2 \pi}{\omega}
-     */
-    virtual double expectedPointCloudNum(double distance, double area) const
-    {
-      assert(spindle_velocity_ != 0.0);
-      double dphi = 2.0 * M_PI / spindle_velocity_;
-      return point_sample_ / (2.0 * M_PI * dphi) / (distance * distance) * area;
+    else {  
+      boost::normal_distribution<> dst(mean, sqrt(var));
+      boost::variate_generator<
+        boost::mt19937&,
+        boost::normal_distribution<> > rand(gen, dst);
+      return rand();
     }
-    
-  protected:
-    
-    double spindle_velocity_;
-    double min_angle_;
-    double max_angle_;
-    size_t point_sample_;
-  private:
-    
-  };
+  }
+  
+  double randomUniform(double min, double max, boost::mt19937& gen)
+  {
+    // Ensure min < max
+    double amin = std::min(min, max);
+    double amax = std::max(min, max);
+    boost::uniform_real<> dst(amin, amax);
+    boost::variate_generator<
+      boost::mt19937&,
+      boost::uniform_real<> > rand(gen, dst);
+    return rand();
+  }
 }
-
-#endif
